+++ base/deprecated.jl
@@ -897,10 +897,11 @@ end
 
 # issue #24794
 @deprecate linspace(start, stop)     linrange(start, stop, 50)
-@deprecate logspace(start, stop)     logspace(start, stop, 50)
+@deprecate logspace(start, stop)     logrange(start, stop, 50)
 
 @deprecate linspace linrange
 @deprecate_binding LinSpace LinRange
+@deprecate logspace logrange
 
 @deprecate merge!(repo::LibGit2.GitRepo, args...; kwargs...) LibGit2.merge!(repo, args...; kwargs...)
 @deprecate push!(w::LibGit2.GitRevWalker, arg) LibGit2.push!(w, arg)